Fix Bug - Appointment with empty interviewer - edit crash
when an appointment was saved without an interviewer and subsequently edited the form would crash due to the null interviewer.  Added validation on the form save to require an interviewer to be selected and added interviewer null check to allow an appointment with null interviewer to be edited incase an appointment with null interviewer still existed in the system.